+++ test/check-error-codes.py
@@ -41,7 +41,8 @@ known_untested_codes = {
     "M0164", # unknown record or variant label in textual representation
     "M0165", # odd expected type
     "M0191", # compiler warning about wasm features (hard to trigger)
-    "M0225", # wasm components support not enabled, use -wasm-components flag
+    "M0232", # cannot infer type of implicit argument
+    "M0234", # wasm components support not enabled, use -wasm-components flag
     }
 
 def populate_error_codes():