+++ Changelog.md
@@ -1,5 +1,13 @@
 # Motoko compiler changelog
 
+## Unreleased
+
+* motoko (`moc`)
+
+  * Experimental support for `implicit` argument declarations (#5517).
+
+  * Experimental support for Mixins (#5459).
+
 ## 0.16.3 (2025-09-29)
 
 * motoko (`moc`)